+++ src/logic/evaluation.ts
@@ -6,7 +6,7 @@ import { Scope } from './scope'
 import { TypeChecker } from './typeChecker'
 import { Substitution } from './typeUnifier'
 import { Reporter } from '../utils/reporter'
-import { declarationPathTo } from '../helpers/logicAst'
+import { declarationPathTo } from './ast'
 import { assertNever } from '../utils/typeHelpers'
 
 const STANDARD_LIBRARY = 'standard library'
